==> 00_IntegratingDatasets/03_a_CreateAnnotationDatabase.R <==
library(tidyverse)
projectDir <- "~/BreastCancerG0arrest"
setwd("00_IntegratingDatasets")

# Prepare cell type database----
## Download & Read data----
dir.create("database", recursive = F)

curl::curl_download("https://panglaodb.se/markers/PanglaoDB_markers_27_Mar_2020.tsv.gz", destfile = "IntegratingDatasets/database/PanglaoDB_markers_27_Mar_2020.tsv.gz")

panglao.db <- read.csv("IntegratingDatasets/database/PanglaoDB_markers_27_Mar_2020.tsv.gz", stringsAsFactors = F, header = T, sep = "\t")

## Filter-out mouse-only signatures----
panglao.db <- panglao.db[panglao.db$species != "Mm", -c(4:9, 11:14)]

## Remove unnecessary data----
panglao.db$species <- panglao.db$organ
panglao.db <- panglao.db[ , 1:3]
colnames(panglao.db) <- c("tissueType", "geneSymbolmore1", "cellName")

panglao.db$tissueType[is.na(panglao.db$tissueType)] <- "Other"

## Keep only tissues of interest----
panglao.db <- panglao.db[(panglao.db$tissueType == "Immune system" | 
                            panglao.db$tissueType == "Connective tissue" |
                            panglao.db$tissueType == "Epithelium" |
                            panglao.db$tissueType == "Vasculature" |
                            panglao.db$tissueType == "Blood" |
                            panglao.db$tissueType == "Mammary gland") |
                            panglao.db$cellName == "Vascular smooth muscle cells" |
                            panglao.db$cellName == "T cells naive", ]

## Remove redundant cell types----
unique(panglao.db$cellName)

to_remove <- c("Endothelial cells (aorta)", "Endothelial cells (blood brain barrier)","Chondrocytes", "Erythroblasts", "Red pulp macrophages", "Erythroid-like and erythroid precursor cells")

for (remove in to_remove) {panglao.db <- subset(panglao.db, cellName != remove)}

## Replace the first column with the tissue of interest----
panglao.db$tissueType <- "breast"
types <- unique(panglao.db$cellName)

db <- data.frame()

## Extract gene names and place them next to each cell type as comma separated
for (celltype in types) {
  genes <- panglao.db[panglao.db$cellName == celltype, 2]
  row <- cbind("breast", celltype, paste(as.character(genes), sep="' '", collapse=","))
  db <- rbind(db, row)
}

# two more columns are required by the annotation script:
db$geneSymbolmore2 <- NA; db$shortName <- NA

## rename db column names
colnames(db) <- c("tissueType", "cellName", "geneSymbolmore1", "geneSymbolmore2", "shortName")

## If there are known down-regulated gene information for a particular cell type, place them in geneSymbolmore2
db$geneSymbolmore2[db$cellName == "Platelets"] <- "PECAM1,SPN,VEGFA"
db$geneSymbolmore2[db$cellName == "Adipocyte progenitor cells"] <- "CEBPA,GLI1,SULT1E1,EBF2,NRG4"
db$geneSymbolmore2[db$cellName == "Adipocytes"] <- "SLC36A2,LPIN1,APPL2,APOL6"
db$geneSymbolmore2[db$cellName == "Fibroblasts"] <- "IL1R1,FLI1,FABP4,COPZ2,FOSB,NFATC1"
db$geneSymbolmore2[db$cellName == "Basal cells"] <- "DUSP23,MMP7"
db$geneSymbolmore2[db$cellName == "Stromal cells"] <- "SNED1,PDGFRB,WNT2"
db$geneSymbolmore2[db$cellName == "Epithelial cells"] <- "KRT7"
db$geneSymbolmore2[db$cellName == "Mesothelial cells"] <- "OSR1,TWSG1"
db$geneSymbolmore2[db$cellName == "B cells"] <- "CD2,CD80,LTB,MME,IGHG1"
db$geneSymbolmore2[db$cellName == "Basophils"] <- "GPR97,LAMA5,CDH1,HGF"
db$geneSymbolmore2[db$cellName == "Dendritic cells"] <- "HLA-DQA1,CCL22,AP1S3,FCER1A"
db$geneSymbolmore2[db$cellName == "Eosinophils"] <- "IL5RA,CCL11,PRG3,SIGLEC10,S100A9,FUT4"
db$geneSymbolmore2[db$cellName == "Gamma delta T cells"] <- "MKI67,TRGV10,TRGV2,PCNA"
db$geneSymbolmore2[db$cellName == "Macrophages"] <- "LILRB4,CX3CR1,SLAMF7,SLC7A7"
db$geneSymbolmore2[db$cellName == "Mast cells"] <- "BMPR2,DCLK3,HNMT"
db$geneSymbolmore2[db$cellName == "Monocytes"] <- "RGS1,FCGR3B,PADI4,GHSR,CD36"
db$geneSymbolmore2[db$cellName == "Neutrophils"] <- "SERPINB1"
db$geneSymbolmore2[db$cellName == "Nuocytes"] <- "RORA"
db$geneSymbolmore2[db$cellName == "Plasma cells"] <- "P2RX5"
db$geneSymbolmore2[db$cellName == "Plasmacytoid dendritic cells"] <- "CLEC4G,TLR7"
db$geneSymbolmore2[db$cellName == "T cells"] <- "CCL3,CD4,CD7,CCR7"
db$geneSymbolmore2[db$cellName == "T cytotoxic cells"] <- "CD28"
db$geneSymbolmore2[db$cellName == "T follicular helper cells"] <- "TNFSF4"
db$geneSymbolmore2[db$cellName == "T helper cells"] <- "AHR"
db$geneSymbolmore2[db$cellName == "T memory cells"] <- "ARHGEF1"
db$geneSymbolmore2[db$cellName == "Endothelial cells"] <- "GPIHBP1,KLK1,ADORA2A,ARAP3"
db$geneSymbolmore2[db$cellName == "Pericytes"] <- "GPIHBP1,EMID1,SLC38A11,FABP4,CYGB,ASPN,REM1,RARRES2,SERPING1,ECM1"

db$geneSymbolmore1[db$cellName == "Basal cells"] <- "KRT14,KRT5,KRT17,SFN,ACTG2,MYLK,SAA1,C2orf40,DST,LAMB3,SERPINB5,MIA,NEDD4L,ACTA2,CNN1,OXTR,STK17A,AMIGO2,IRF6,S100A2,CRYAB,SYNM,TES,SPP1,SYT8,MYH11,FBXO2,PPP1R14A,LAMC2,ARHGEF35,ADRB2,COL17A1,HBEGF,LAMA3,SGK1,TNS4,TPM1,FBXO32,TAGLN,ANO1,BEX2,WIF1,ACTN1,TFAP2C,GAN,CALML3,CDH3,ACTN4,SEMA3C,CD200,IL17B,C9orf3,SLMAP,IRX1,GRP,DKK3,PDLIM3,FAM126A,PTP4A1,TUBB2B,FHL2,CSRP1,IGFBP2,TPM2,PDLIM4,ANXA1,MT1X,TMEM40,MTHFD2,KLC3,TRNP1,PLP2,CYCS,NFIB,TRIM29,MIR22HG,SLC1A5,ID4,FOSL1,PLK2,GPR87,KCNMB1,SLC3A2,ZNF503,ATF4,ITGB1,MYL9,TUBB2A,RIPK4,STAC2,STMN1,SFPQ,H2AFZ,DUSP14,NGFR,PTPRE,TMEM30B,NTF3,HOTAIRM1,TIPARP,ETF1,MFSD2A,PTN,NOP16,TUBB4B,TGIF1,NSG1,EPOP,COL9A2,GNL3,RASSF10,NCL,MYC,NOLC1,EIF4G2,WDR43,EIF4A1,GADD45A,TRIB1,SNCA,SERBP1,EIF3J,NFASC,RNF152,CARMN,TLE4,HNRNPAB,MAFF,LRRC8A,FAM189A2,HNRNPA2B1,DOK5,DNTTIP2,PPP1R13L,ARHGEF5,HSPA2,AMD1,UBE2S,MT1G,LIMK2,HOMER1,FST,NETO2,MT1F,CLTB,TCEAL2,SLC25A4,MAOB,IRX2,HEXIM1,PNO1,SBDS,HNRNPM,CSTA,ARC,MME,NT5DC3,PALLD,DDX5,SMTN,HRAS,SLC2A1,RTN4,LY6K,KIF5B,TXNRD1,BCAR3,HSP90AB1,ERF,IFFO2,P3H2,MAT2A,RGS12,TPM4,HNRNPH1,CD3EAP,EIF5,IFRD1,LPCAT2,UTP4,GRPEL1,TMEM237,NES,PPP1R12A,TLE3,DEGS1,POLR1C,SCPEP1,NEXN,SH3BGR,PSMG1,MGME1,RHEB,RPS6KA5,FOXC2,NXF1,PTPN14,GCLM,DRAP1,THBS1,CTNNB1,ITGA10,PTS,ALDH7A1,RBBP8,TNC,EIF4E,DCAF13,ARL6IP1,BLCAP,MRPL32,NOP58,USP31,KLHL2,KHDRBS3,RRS1,PRNP,GPATCH4,TPI1,MAPRE1,MLF1,NIP7,LBH,NNMT,MED21,GRWD1,HNRNPU,BRIX1,GRAMD2B,UBR4,RRAS2,CRISPLD1,NOP56,PXDC1,DYNLL1,H2AFX,MT1E,UBE2N,ANP32E,SMURF2,CNN3,DLL1,HNRNPH3,SYNCRIP,KDSR,SRSF8,WLS,TXLNG,ZMAT3,LTBP2,APOE,CC2D1A,CMC2,LIMA1,ESF1,PPP1R12B,PAMR1,USP1,MATN2,TRMT10C,KLHL13,SON,TFB2M,BIRC2,CHIC2,TDRP,FJX1,MXI1,SOWAHC,CCDC86,BYSL,C1orf109,ETNK1,NOL10,SMC5,ARHGAP5,RNF103,SARS,POLE3,SBNO1,NUDT4,TSR1,HSPA9,ING2,CTR9,NXT1,EIF4A3,TOB2,METAP2,NDEL1,UBIAD1,ARIH1,SAFB,UTP11,PITHD1,TOP1,SEH1L,BRD2,RCL1,NFAT5,PER2,YWHAG,ZNF207,CEBPZ,MIR222HG,COPS3,RBM25,PTDSS1,MED27,SNX18,POSTN,CLP1,SLC41A1,GPC1,EXOSC10,NOM1,PNRC2,FRMD6,FAM210A,PRPF4B,SRSF1,TFAM,MEPCE,CNN2,CDC5L,CUL1,ZNF800,ITGBL1,KRR1,ADNP,TRIM28,ZYX,GAS6,ARL13B,HIBCH,PAK1IP1,SBSPON,TGFB1I1,UTP3,USP32,C16orf72,ZNF574,MPHOSPH10,EGR3,RETREG1,NCKAP5L,TEAD3,FZD7"
alveolar <- c(tissueType = "breast", cellName = "Alveolar", geneSymbolmore1 = "PIGR,SLPI,CCL28,KRT23,LTF,CLDN3,ALDH1A3,SNORC,MMP7,GABRP,CDC42EP1,PDZK1IP1,KRT15,PPP1R1B,S100A9,CRABP1,ELF5,S100A1,LCN2,SCGB3A1,ROPN1B,LIF,BBC3,NDRG2,PHLDA2,RCAN1,YWHAH,HMGA1,SDC4,IQCG,CALML5,PI3,CXCL2,SELENBP1,IGF2BP2,RARRES1,PPA1,KRTCAP3,TTYH1,SLC25A37,PLPP2,EIF4EBP1,MTHFD2L,FAM3B,SPTSSA,SCGB1D2,SH3YL1,LINC01198,DBI,PTRHD1,CD14,FOLR1,DAPP1,SCGB2A2,ANKRD36C,SLC5A6,LIPH,CDC42EP5,FGFBP1,LRRC26,CXCL17,PRELID1,CLDN8,HOMER2,PPP1R14B,ROPN1,NCALD,HEBP2,TCN1,INSR,APRT,CTSV,CYP24A1,CHI3L1,USP53,CXCL3,ATP5F1D,CLDN1,TPD52L1,TMPRSS2,AQP5,MFGE8,HCAR2,PRSS8,KCNN4,CITED4,AC025580.1,CXADR,CASC8,ZNF562,TTC9,FAM3D,GBP2,GALNT1,SAA2,RAB9A,C1QBP,AP000851.1,COBL,TRAP1,QARS,SCGB2A1,OLFM4,TKT,RIPK2,ACTR3B,KRT16,CHI3L2,TM7SF2,NCCRP1,BBOX1,FDCSP,DNPH1,STAP2,PPP1R14C,PART1,C19orf70,PLEKHB1,DSC2,TIAM1,KIT,ABHD14B,MYEOV,PADI2,SDCBP2,DNTTIP1,IMPDH2,PODXL2,LMTK3,SLC34A2,SESTD1,KRT7,SLC6A14,AFDN,PROM1,GPAT3,CA8,SLC13A2,SFRP1,TFCP2L1,POLD2,NBEAL1,PLEKHS1,CDCP1,TATDN1,SLC12A2,NDUFV1,SLC5A1,CEACAM1,MRPL15,ITGB8,TMEM213,COX5A,KLK10,EIF3D,DSG2,TRPM4,KLK5,ATP6V1B1,MESP1,PACSIN3,NT5DC2,NPM3,PLEC,FABP3,POU2F3,CD82,PPP1R1A,PYROXD2,FABP7,SLC28A3,CHCHD10,TIMM9,RTP4,SMARCD2,AC009041.2,MELTF,KLK7,TMEM159,C19orf48,CCDC9B,EXOSC5,AC011247.1,ZNF385C,TP53,KARS,STK40,HIST2H2BE,RASL10A,AC004130.1,AL391832.2,TRIM2,ZNF33B,ROCR,DTNB,RAB24,VPS51,EIF4B,ATP13A5,PHB2,CCT8,HSD17B2,PDCD5,ITPKC,PDLIM5,SNRPA,TOP1MT,FAHD2A,SLC20A2,EIF2S3,SLC27A5,EIF3G,CGREF1,EIF3M,BOP1,SRPK1,CCNB1IP1,F12,OBP2B,ETHE1,MRPL45,CYB5R2,KRT6B,FOLH1,ETV6,NOA1,CFL2,ACAT2,PMPCB,CCT3,MAP7D1,MFSD14C,PHGDH,HIST3H2A,IFT172,C5orf46,TEN1,PPP1R16A,SYCE1L,TMEM189,SERTAD2,SOX9,C2orf88,EIF2D,LUCAT1,ABI1,RTKN,C12orf45,MRPS24,AC106869.1,PAICS,SND1,PFDN4,TLE2,NME2,URGCP,NIPAL1,KLK6,LGALSL", geneSymbolmore2 = NA, shortName = NA)
hormone_sensing <- c(tissueType = "breast", cellName = "Hormone sensing", geneSymbolmore1 = "TCIM,AGR3,TFF3,AZGP1,AGR2,ANKRD30A,TFF1,BATF,TMC5,C15orf48,MLPH,EFHD1,STC2,FBP1,ZG16B,SQOR,LURAP1L,KIAA1324,SPDEF,ACADSB,BAMBI,CIB1,RAB11FIP1,SMIM14,GJC3,PKIB,RASEF,MUC1,SCCPDH,ALB,AFF3,RBM47,SYTL2,FOXA1,FASN,SLC40A1,SCUBE2,AR,BSPRY,PIP,VSIG2,CA12,STARD10,TSPAN15,FAM102A,MAL2,CAPG,H2AFJ,CYB5A,NUCB2,S100A14,TSPAN13,TPD52,WFDC2,SLC39A6,ELOVL5,AREG,PTP4A2,XBP1,VMP1,MAGED2,HIGD1A,LIMCH1,ALCAM,TSPAN6,RERG,REEP5,DHCR24,VTCN1,PRDX3,CD164,MYO6,QSOX1,HSBP1,TTC39A,LYPD6B,GPR160,SDF2L1,SLC7A2,SERHL2,PRLR,GOLM1,ERBB3,BIK,C16orf89,VAMP8,PLGRKT,TMEM45B,IL20RA,TFAP2A,INPP4B,CITED1,PHGR1,CAPN8,AK5,REEP6,FTO,CRB3,HMGN1,COX6C,KRT10,SLC9A3R1,PTHLH,TMEM160,C3orf14,SRI,CCDC57,TOX3,LAMTOR2,MISP,TOM1L2,TBX3,SRARP,MAP3K1,TM4SF1-AS1,NEK10,RAB27B,ERBB2,MREG,ADAM10,COA3,HMGCS2,PLEKHF2,PSENEN,DIO2,SULT2B1,NQO1,SEZ6L2,ERBB4,GALNT6,DEGS2,VAV3,CNDP2,MPHOSPH6,SHROOM3,TRAFD1,FAM174B,SLC44A4,KCNK6,CMBL,GMPPB,MEAF6,PGM2L1,ZKSCAN1,SHISA2,ERG28,GLUD1,IBTK,TMCO1,ESR1,DICER1,ACADS,TXNDC17,SLC7A8,SH3BGRL,AC097059.1,EGOT,CISD3,SMAD3,KIF12,SEMA4B,HIPK2,MUCL1,FLNB,MB,SUSD3,AIG1,TSPAN5,SLC26A3,DCDC2,SYTL4,IRS1,AKR7A3,UPK3A,MYB,RNF181,TMEM241,MAPK13,ARFGEF3,ARHGAP8,GDE1,JPT2,HACD3,AP2B1,TMCO3,ALDH3B2,MYBPC1,MRPS10,ITGAV,CDH1,SPINT2,SDR16C5,INPP5J,EEF1A2,CERS4,C3orf52,DPY30,AL121944.1,SH3BP4,SMIM29,SGK3,EIF2AK1,TMEM54,CHCHD5,SEMA3E,G6PC3,TTC36,WBP1L,SLC2A10,RUNX1,BAG1,C9orf152,WEE1,RBP1,HEPACAM2,HID1,TMEM136,PDZK1,VSIG10L2,AL049839.2,COPS9,FAAH,TCEAL3,NDUFA5,MPP7,ARHGAP32,SUSD6,PANK3,LRIG1,LINC00504,DLG5,ACAA1,C7orf50,RABEP1,ZNF292,CNTD1,GREB1,SRP9,HPX,LRP2,CPEB4,XAB2,AC093001.1,EFHC1,ZNF91,ATP6V0B,HSD17B4,OXLD1,CCDC170,CHPF,ARSD,FAM234B,SHTN1,ADGRV1,PARD6B,GTF3C6,SYT7,KIAA1522,HERC4,SLC4A7,CETN2,KDM4B,SLC25A39,PLOD2,AC004982.2,PTPRK,USP22,KIF16B,IRX3,SPR,TP53TG1,GALNT7,PGM3,LIN7A,RAB17,FMO5,MAGEF1,ETFA,ZNF165,TMEM254,BLVRB,IDH1,PYM1,ARSA,TBC1D9,C12orf49,CYB5R1,SNRPD3,SPTSSB,CTNND1,SCAMP4,TRPS1,MYCBP,SLC35A3,UGDH,ACLY,INTS12,CCDC159,CTXN1,LINC00467,AC008771.1,ENPP5,DDX17,BCAS4,ENTPD5,RUNDC1,ZNF652,CLDN12,PRKAA2,GMNN,PARP1,AQP3,RALGPS2,CCDC74A,CLTC,PLAT,LIPE-AS1,MBTPS1,TFAP2B,TMEM60,ABCD3,ZBED5-AS1,SORD,CADPS2,PLEKHG3,NRIP1,CRNDE,WWC1,INSIG2,SLC22A18,SORT1,GATA3-AS1,EPB41L4B,PPP1R35,GTF2I,LZTFL1,ENPP1,GMPPA,SMIM15,TP53I11,PGR,CARD19,AAGAB,NDUFAF6,CNFN,BLVRA,GPRC5C,KIAA0319L,RHPN1,TESMIN,TMED7,SCD,TSPAN1,FAM174A,LPGAT1,LNX2,CD2AP,TMEM87B,GFPT1,WNT4,LRBA,SPA17,MBOAT7,AL031708.1,RNF141,CPEB2,GPATCH2,IGF1R,LMO7,MSI2,SPOPL,ARF3,ZNF626,SPAG1,MLEC,NTN4,DUSP10,VIRMA,TMEM25,AL355338.1,CMTM4,PPP1R37,TOR1AIP2,SCOC,GTF2H2,IFT122,ATF6,RHOD,ACSL3,PFN2,THSD4,MAGI3,ACSF2,SEC14L2,MEGF9,EBF4,SEC16A,FAM120A,MSRB1,SLC22A5,HSBP1L1,MAPK8,PTPN3,MYOF,RAP1GDS1,C19orf33,DAXX,ZNF664,PDZD8,TNIK,POP7,FGFR1OP,CD2BP2,RTN2,IL13RA1,ZNF385A,NCEH1,VEZT,TMEM41B,RNF39", geneSymbolmore2 = NA, shortName = NA)
db <- rbind(db, alveolar, hormone_sensing)

saveRDS(db, "./rds/db.rds")

sessionInfo()
